+++ packages/typespec-rust/src/codegen/context.ts
@@ -86,6 +86,7 @@ export class Context {
         }
 
         switch (method.returns.type.kind) {
+          case 'pageIterator':
           case 'pager': {
             recursiveAddBodyFormat(method.returns.type.type.content, helpers.convertResponseFormat(method.returns.type.type.format));
             break;