+++ packages/typespec-rust/src/tcgcadapter/adapter.ts
@@ -2332,9 +2332,24 @@ export class Adapter {
         throw new AdapterError('InternalError', `failed to unwrap paged items for method ${method.name}`, method.__raw?.node);
       }
 
-      this.crate.addDependency(new rust.CrateDependency('async-trait'));
+      const pagedResponseType = new rust.Response(this.crate, synthesizedModel, responseFormat);
+
       // default to nextLink. will update it as required when we have that info
-      rustMethod.returns = new rust.Result(this.crate, new rust.Pager(this.crate, new rust.Response(this.crate, synthesizedModel, responseFormat), 'nextLink'));
+      const continuationKind: rust.PagerContinuationKind = 'nextLink';
+
+      // if the method has the forcePageIterator client option set to true,
+      // expose this as a PageIterator so all items are accessible.
+      const forcePageIterator = method.decorators.find(
+        d => d.name === 'Azure.ClientGenerator.Core.@clientOption' && d.arguments['name'] === 'forcePageIterator'
+      )?.arguments['value'] as boolean | undefined;
+      let resultType: rust.PageIterator | rust.Pager;
+      if (forcePageIterator) {
+        resultType = new rust.PageIterator(this.crate, pagedResponseType, continuationKind);
+      } else {
+        this.crate.addDependency(new rust.CrateDependency('async-trait'));
+        resultType = new rust.Pager(this.crate, pagedResponseType, continuationKind);
+      }
+      rustMethod.returns = new rust.Result(this.crate, resultType);
     } else if (method.kind === 'lro') {
       const pushModels = (
         tcgcType: tcgc.SdkType,
@@ -2524,6 +2539,7 @@ export class Adapter {
     // response header traits are only ever for marker types and payloads
     let implFor: rust.AsyncResponse<rust.MarkerType> | rust.Response<rust.MarkerType | rust.Model>;
     switch (method.returns.type.kind) {
+      case 'pageIterator':
       case 'pager':
       case 'poller':
         implFor = method.returns.type.type;